+++ crates/core/src/lib.rs
@@ -136,12 +136,13 @@ pub mod openapi {
   pub struct Doc;
 }
 
+
 pub mod api {
   pub use crate::admin::user::{CreateUserRequest, create_user_handler};
   pub use crate::auth::{AuthTokenClaims, JwtHelper, cli};
   pub use crate::connection::Connection;
   pub use crate::email::{Email, EmailError};
-  pub use crate::migrations::new_unique_migration_filename;
+  pub use crate::migrations::{apply_declarative_schema, new_unique_migration_filename};
   pub use crate::records::json_schema::build_api_json_schema;
   pub use crate::schema_metadata::ConnectionMetadata;
   pub use crate::server::{InitArgs, init_app_state, serve};